% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_notice.R
\name{.ukbrapr_startup_notice}
\alias{.ukbrapr_startup_notice}
\title{Package notice}
\usage{
.ukbrapr_startup_notice()
}
\value{
Invisible NULL.
}
\description{
Prints a feedback notice once per session.
}
\keyword{internal}
