% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_rap.R
\name{download_from_rap}
\alias{download_from_rap}
\title{Use R to download a file to the UK Biobank RAP}
\usage{
download_from_rap(file, verbose = FALSE)
}
\arguments{
\item{file}{A string. Filename of the file to be downloaded (character)}

\item{verbose}{Logical. Be verbose,
\code{default=FALSE}}
}
\value{
NA
}
\description{
Use R to download a file to the UK Biobank RAP (really just a wrapper for `dx download`)
}
\examples{

readr::write_tsv(data.frame(x=1:10,y=11:20), "ukbrap.dummy.20231114.txt.gz")

# download file to RAP storage
download_from_rap(file="ukbrap.dummy.20231114.txt.gz", dir="extracts/")

}
\author{
Luke Pilling
}
