% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{export_tables_baseline_dates}
\alias{export_tables_baseline_dates}
\title{Extract useful baseline date fields and save to file}
\usage{
export_tables_baseline_dates(dataset = NULL, submit = FALSE, verbose = FALSE)
}
\value{
NA
}
\description{
In the UK Biobank RAP, submit a table-exporter job to extract some useful baseline fields:

1) date of baseline assessment [p53_i*], 2) month of birth [p52], 3) year of birth [p34], and 4) sex [p31]
}
\author{
Luke Pilling
}
