% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{export_tables_cancer_registry}
\alias{export_tables_cancer_registry}
\title{Extract cancer registry fields}
\usage{
export_tables_cancer_registry(
  n_cancer_arrays = 21,
  dataset = NULL,
  submit = FALSE,
  verbose = FALSE
)
}
\value{
NA
}
\description{
In the UK Biobank RAP, submit a table-exporter job to extract the cancer registry fields:
}
\author{
Luke Pilling
}
