% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{export_tables_emr}
\alias{export_tables_emr}
\title{Export tables for HES, GP, and death}
\usage{
export_tables_emr(dataset = NULL, submit = FALSE, verbose = FALSE)
}
\value{
NA
}
\description{
In the UK Biobank RAP export tables for HES, GP, and death using the table-exporter
}
\author{
Luke Pilling
}
