% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{export_tables_selfrep_illness}
\alias{export_tables_selfrep_illness}
\title{Extract self-reported illess fields}
\usage{
export_tables_selfrep_illness(
  n_cancer_arrays = 5,
  n_noncancer_arrays = 30,
  n_instances = 3,
  dataset = NULL,
  submit = FALSE,
  verbose = FALSE
)
}
\value{
NA
}
\description{
In the UK Biobank RAP, submit a table-exporter job to extract the self-reported illness fields:
}
\author{
Luke Pilling
}
