% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cancer_registry.R
\name{get_cancer_registry}
\alias{get_cancer_registry}
\title{Get UK Biobank participant self-reported illness/year data for specific codes}
\usage{
get_cancer_registry(codes, ukb_dat, verbose = FALSE)
}
\arguments{
\item{codes}{A vector. The ICD10 codes to ascertain cancer registry entires}

\item{ukb_dat}{A data frame. Contains the participant-level cancer registry fields e.g., `p40005_i0`}

\item{verbose}{Logical. Be verbose,
\code{default=FALSE}}
}
\value{
Returns a data frame with four variables: eid, selfrep [binary, codes identified?], selfrep_df [date of reported illness], selfrep_i [instance the illness was first reported]
}
\description{
For a specific self-reported illness code or codes, identify whether the participant has self-reported at any visit, and identify the year. 
Intended for use on the UK Biobank DNnexus Research Analysis Platform, but if the user provides their own dataframe of UK Biobank self-reported fields this works on any system.
}
\examples{
# example diagnostic codes for haemochromatosis
print(codes_df_hh)

# get self-reported data - a data frame
selfrep_df <- get_selfrep_illness(codes_df_hh)

# inspect variables
table(selfrep_df$selfrep)
summary(selfrep_df$selfrep_df)
table(selfrep_df$selfrep_i)

}
\author{
Luke Pilling
}
