% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emr_diagnoses.R
\name{get_emr_diagnoses}
\alias{get_emr_diagnoses}
\title{Get UK Biobank participant Electronic Medical Records (EMR) data}
\usage{
get_emr_diagnoses(
  codes_df,
  vocab_col = "vocab_id",
  codes_col = "code",
  spark_master = "spark://master:41000",
  verbose = FALSE
)
}
\arguments{
\item{codes_df}{A data frame. Contains the `vocab_col` and `codes_col` i.e., a list of diagnostic codes, and an indicator of the vocabulary.}

\item{vocab_col}{A string. Column name in `codes_df` that contains the vocabulary indicator for the code (vocab IDs should be 'ICD10' 'Read2' or 'CTV3').
\code{default='vocab_id'}}

\item{codes_col}{A string. Column name in `codes_df` that contains the diagnostic code ('ICD10' 'Read2' or 'CTV3' codes).
\code{default='code'}}

\item{spark_master}{A string. The `master` argmuent passed to `sparklyr::spark_connect()`.
\code{default='spark://master:41000'}}

\item{verbose}{Logical. Be verbose,
\code{default=FALSE}}
}
\value{
Returns a list of data frames (the participant data for the requested diagnosis codes: `death_cause`, `hesin_diag`, and `gp_clinical`. Also includes the original codes list)
}
\description{
Using a Spark node/cluster on the UK Biobank Research Analysis Platform (DNAnexus), use R to get medical records for specific diagnostic codes list
}
\examples{
# example diagnostic codes for CKD from GEMINI multimorbidity project
codes_df <- readr::read_tsv("https://raw.githubusercontent.com/GEMINI-multimorbidity/diagnostic_codes/main/codelists/CKD.txt")
codes_df

# get diagnosis data - returns list of data frames (one per source)
diagnoses_list <- ukbrapR::get_emr_diagnoses(codes_df)
# 7 ICD10 codes, 40 Read2 codes, 37 CTV3 codes 
# 298.18 sec elapsed

# save to files on the RAP worker node
readr::write_tsv(diagnoses_list$death_cause, "ukb14631.CKD.death_cause.20231114.txt.gz")
readr::write_tsv(diagnoses_list$hesin_diag,  "ukb14631.CKD.hesin_diag.20231114.txt.gz")
readr::write_tsv(diagnoses_list$gp_clinical, "ukb14631.CKD.gp_clinical.20231114.txt.gz")

# upload data to RAP storage
ukbrapR::upload_to_rap(file="ukb14631.CKD.*.20231114.txt.gz", dir="")

}
\author{
Luke Pilling
}
