% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emr_local.R
\name{get_emr_local}
\alias{get_emr_local}
\title{Get UK Biobank participant Electronic Medical Records (EMR) data}
\usage{
get_emr_local(
  codes_df,
  vocab_col = "vocab_id",
  codes_col = "code",
  local_paths = NULL,
  verbose = FALSE
)
}
\arguments{
\item{codes_df}{A data frame. Contains the `vocab_col` and `codes_col` i.e., a list of diagnostic codes, and an indicator of the vocabulary.}

\item{vocab_col}{A string. Column name in `codes_df` that contains the vocabulary indicator for the code (vocab IDs should be 'ICD10' 'Read2' or 'CTV3').
\code{default='vocab_id'}}

\item{codes_col}{A string. Column name in `codes_df` that contains the diagnostic code ('ICD10' 'Read2' or 'CTV3' codes).
\code{default='code'}}

\item{local_paths}{A data frame. Columns must be `object` and `path` containing paths to: 
`death`, `death_cause`, `hesin`, `hesin_diag` & `gp_clinical` 
\code{default=NULL}}

\item{verbose}{Logical. Be verbose,
\code{default=FALSE}}
}
\value{
Returns a list of data frames (the participant data for the requested diagnosis codes: `death_cause`, `hesin_diag`, and `gp_clinical`. Also includes the original codes list)
}
\description{
Get medical records for specific diagnostic codes list using local UK Biobank EMR long files
}
\examples{
# example diagnostic codes for CKD from GEMINI multimorbidity project
head(codes_df_ckd)

# get EMR data - returns list of data frames (one per source)
emr_dat <- get_emr_local(codes_df_ckd)

# save to data either as an R object, or separate as text files:
save(emr_dat, "ukbrap.CKD.emr.20231114.RDat")
readr::write_tsv(emr_dat$hesin_diag,  "ukbrap.CKD.hesin_diag.20231114.txt.gz")

}
\author{
Luke Pilling
}
