% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emr_spark.R
\name{get_emr_spark}
\alias{get_emr_spark}
\title{Get UK Biobank participant Electronic Medical Records (EMR) data in a RAP Spark environment}
\usage{
get_emr_spark(
  codes_df = NULL,
  spark_master = "spark://master:41000",
  verbose = FALSE
)
}
\arguments{
\item{codes_df}{A data frame. Contains two columns: `code` and `vocab_id` i.e., a list of diagnostic codes, and an indicator of the vocabulary. Other columns are ignored.}

\item{spark_master}{A string. The `master` argmuent passed to `sparklyr::spark_connect()`.
\code{default='spark://master:41000'}}

\item{verbose}{Logical. Be verbose,
\code{default=FALSE}}
}
\value{
Returns nothing
}
\description{
This function is completely removed. Better to use `get_diagnoses()`. Use a historic release of this package if you really need it https://github.com/lcpilling/ukbrapR/releases
}
\author{
Luke Pilling
}
