% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imputed_variant_info.R
\name{get_imputed_variant_info}
\alias{get_imputed_variant_info}
\title{Get variant info from UK Biobank imputed genotype MFI files}
\usage{
get_imputed_variant_info(varlist, verbose = FALSE)
}
\arguments{
\item{varlist}{A data frame. Contains at least two columns: `chr` and `pos` (b37). Other columns are ignored.}

\item{verbose}{Logical. Be verbose (show individual steps),
\code{default=FALSE}}
}
\value{
A data frame of variants with added "ukb_rsid" column (also returns alleles, MAF and INFO)
}
\description{
For a given set of genomic coordinates (position in build 37) get the UK Biobank imputed genotype variant IDs from the MFI files.
}
\examples{

# Example where RSID is only known for some variants
varlist1 <- data.frame(rsid=c("","rs564086017",""), chr=c(1,2,22), pos=c(10616, 10180, 16050435))
varlist2 <- get_imputed_variant_info(varlist1)

}
\author{
Luke Pilling
}
