% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phenos_rap_spark.R
\name{get_rap_phenos}
\alias{get_rap_phenos}
\title{Use R on the UK Biobank RAP (Spark cluster) to get phenotype data}
\usage{
get_rap_phenos(names, record = NULL, verbose = FALSE)
}
\arguments{
\item{names}{A string or vector of strings. The variable name(s) required. e.g., c("eid","p31","p21003_i0") (character)}

\item{record}{A string. The `dnanexus_link` file descriptor of the .dataset to use. Default (if left as NULL) is to use the most recent update (character)}

\item{verbose}{Logical. Be verbose,
\code{default=FALSE}}
}
\value{
Returns a data.frame (the participant data for the requested variables)
}
\description{
Using a Spark node/cluster on the UK Biobank RAP, run R to extract a provided set of variables. Using code from the UK Biobank DNAnexus team https://github.com/UK-Biobank/UKB-RAP-Notebooks/blob/main/NBs_Prelim/105_export_participant_data_to_r.ipynb
}
\examples{
# get phenotype data
df <- get_rap_phenos(c("eid","p31","p21003_i0"))

# save to file on the RAP worker node
write_tsv(df, "ukb14631.data_output.20231026.txt.gz")

# upload data to RAP storage
ukbrapR::upload_to_rap(file="ukb14631.data_output.20231026.txt.gz", dir="")

}
\author{
Luke Pilling
}
