% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bed.R
\name{load_bed}
\alias{load_bed}
\title{Load BED file into memory}
\usage{
load_bed(in_bed, verbose = FALSE, very_verbose = FALSE)
}
\arguments{
\item{in_bed}{A string. BED prefix}

\item{verbose}{Logical. Be verbose (show individual steps),
\code{default=FALSE}}

\item{very_verbose}{Logical. Be very verbose (show individual steps & show terminal output from Plink etc),
\code{default=FALSE}}
}
\value{
A data frame
}
\description{
Use Plink to convert BED to RAW then easily load it
}
\examples{

liver_variants <- load_bed(in_bed="liver_cirrhosis.imputed.variants")

}
\author{
Luke Pilling
}
