% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bed.R
\name{make_imputed_bed}
\alias{make_imputed_bed}
\title{Extract variants from imputed genotype file(s) into single BED file}
\usage{
make_imputed_bed(
  in_file,
  out_bed,
  use_pos = FALSE,
  progress = TRUE,
  verbose = FALSE,
  very_verbose = FALSE
)
}
\arguments{
\item{in_file}{A data frame or file path. Contains at least two columns: (`rsID` and `CHR`) OR (`CHR` and `POS`). Other columns are ignored.}

\item{out_bed}{A string. Output BED file prefix.}

\item{use_pos}{Logical. Use genomic position (CHR and POS) instead of RSID,
\code{default=FALSE}}

\item{progress}{Logical. Show progress through each individual file,
\code{default=TRUE}}

\item{verbose}{Logical. Be verbose (show individual steps),
\code{default=FALSE}}

\item{very_verbose}{Logical. Be very verbose (show individual steps & show terminal output from Plink etc),
\code{default=FALSE}}
}
\value{
A single merged BED file (and BIM and FAM files)
}
\description{
For a given set of identifiers (RSID or "build37 position") extract from the UK Biobank imputed genotypes (BGEN files, field 22828) into a single BED file.
}
\examples{

# Example list of variants
#    only RSID and CHR are provided
varlist <- data.frame(rsid=c("rs1800562","rs429358"), chr=c(6,19))

make_imputed_bed(in_file=varlist, out_bed="example_variants_by_rsid")

# Above example works fine if the imputed BGEN contains the provided rsid

# However if you have a list of variants defined by genomic position (CHR and POS)
# then use the `use_pos=TRUE` argument

# Example list of variants using position (build 37) not rsid
varlist_pos <- data.frame(chr=c(6,19), pos=c(26093141,45411941))

# if you provide an rsid it is ignored when `use_pos=TRUE`
make_imputed_bed(in_file=varlist_pos, out_bed="example_variants_by_pos", use_pos=TRUE)

}
\author{
Luke Pilling
}
