% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_rap.R
\name{upload_to_rap}
\alias{upload_to_rap}
\title{Use R to upload a file to the UK Biobank RAP}
\usage{
upload_to_rap(file, dir = "FALSE", verbose = FALSE)
}
\arguments{
\item{file}{A string. Filename of the file to be uploaded (character)}

\item{dir}{A string. Target directory in the RAP space. If blank, the current working directory (character)}

\item{verbose}{Logical. Be verbose,
\code{default=FALSE}}
}
\value{
NA
}
\description{
Use R to upload a file to the UK Biobank RAP (really just a wrapper for `dx upload`)
}
\examples{
# upload file to RAP storage
ukbrapR::upload_to_rap(file="ukb14631.data_output.20231026.txt.gz", dir="extracts/")

}
\author{
Luke Pilling
}
