% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diagnoses.R
\name{get_diagnoses}
\alias{get_diagnoses}
\title{Get UK Biobank participant Electronic Medical Records (EMR) data}
\usage{
get_diagnoses(codes_df, file_paths = ukbrapR:::ukbrapr_paths, verbose = FALSE)
}
\arguments{
\item{codes_df}{A data frame. Contains two columns: `code` and `vocab_id` i.e., a list of diagnostic codes, and an indicator of the vocabulary. Other columns are ignored.}

\item{file_paths}{A data frame. Columns must be `object` and `path` containing paths to required files. Default assumes you have the tables exported in the RAP environment from
ukbrapR::export_tables() 
\code{default=ukbrapR:::ukbrapr_paths}}

\item{verbose}{Logical. Be verbose,
\code{default=FALSE}}
}
\value{
Returns a list of data frames (the participant data for the requested diagnosis codes: `death_cause`, `hesin_diag`, and `gp_clinical`. Also includes the original codes list)
}
\description{
Get UK Biobank medical records for specific diagnostic codes list

If `file_paths` to exported tables are provided, will extract records from these (i.e., does not need to be launched in a Spark cluster)
}
\examples{
# example diagnostic codes for CKD from GEMINI multimorbidity project
head(codes_df_ckd)

# get EMR data - returns list of data frames (one per source)
# Assumes exported data are in the default RAP locations from `export_tables()` (ukbrapR:::ukbrapr_paths)
emr_dat <- get_diagnoses(codes_df_ckd)

# save to files on the RAP worker node -- either as an R object, or separate as text files:
save(emr_dat, "ukbrap.CKD.emr.20231114.RDat")
readr::write_tsv(emr_dat$hesin_diag,  "ukbrap.CKD.hesin_diag.20231114.txt.gz")

# upload data to RAP storage
upload_to_rap(file="ukbrap.CKD.*", dir="")

}
\author{
Luke Pilling
}
