% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variants.R
\name{make_imputed_bed}
\alias{make_imputed_bed}
\title{Extract variants from imputed genotype file(s) into single BED file}
\usage{
make_imputed_bed(
  in_file,
  out_bed,
  progress = TRUE,
  verbose = FALSE,
  very_verbose = FALSE
)
}
\arguments{
\item{in_file}{A data frame or file path. Contains at least two columns: `rsID` and `CHR`. Other columns are ignored.}

\item{out_bed}{A string.}

\item{progress}{Logical. Show progress through each individual file,
\code{default=TRUE}}

\item{verbose}{Logical. Be verbose (show individual steps),
\code{default=FALSE}}

\item{very_verbose}{Logical. Be very verbose (show individual steps & show terminal output from Plink etc),
\code{default=FALSE}}
}
\value{
A single merged BED file (and BIM and FAM files)
}
\description{
For a given set of genomic coordinates extract from the UK Biobank imputed genotypes (v3) into a single BED file.
}
\examples{

make_imputed_bed(in_file=system.file("files", "pgs_liver_cirrhosis.txt", package="ukbrapR"), out_bed="liver_cirrhosis.imputed.variants")

}
\author{
Luke Pilling
}
